ARTEMIS-47 - Split the tests for each kind of configuration file.
